+++ src/media/audio/pcm/mod.rs
@@ -1,0 +1,21 @@
+// devela::media::audio::pcm
+//
+//! Pulse-code modulation audio buffers and stream metadata.
+//
+
+#[cfg(test)]
+mod tests;
+
+mod buffer; // PcmBuffer
+// mod format; // TODO PcmRaw, PcmWav, …
+mod spec; // PcmSpec, PcmSample
+
+crate::structural_mods! { // _mods
+    _mods {
+        pub use super::{
+            buffer::*,
+            // format::*,
+            spec::*,
+        };
+    }
+}